requirements.txt
# ── Core ML / Embeddings ─────────────────────────────────────────────────────
torch>=2.2.0
transformers>=4.40.0
Pillow>=10.0.0

# ── PDF Parsing ───────────────────────────────────────────────────────────────
PyMuPDF>=1.24.0

# ── Vector Store ─────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── LangChain ────────────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0

# ── LLM API ───────────────────────────────────────────────────────────────────
openai>=1.30.0

# ── Streamlit UI ──────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Environment & Utilities ───────────────────────────────────────────────────
python-dotenv>=1.0.0
numpy>=1.26.0
